@@ -15,11 +15,9 @@
  */
 package com.shakutara.cleanarchitecture.core.platform
 
-import android.arch.lifecycle.ViewModel
-import android.arch.lifecycle.ViewModelProvider
-import android.arch.lifecycle.ViewModelProviders
-import android.support.v4.app.Fragment
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
 import com.shakutara.cleanarchitecture.AndroidApplication
 import com.shakutara.cleanarchitecture.core.di.ApplicationComponent
 import javax.inject.Inject
@@ -39,8 +37,11 @@ abstract class BaseActivity : AppCompatActivity() {
     @Inject
     lateinit var viewModelFactory: ViewModelProvider.Factory
 
-    inline fun <reified T : ViewModel> viewModel(factory: ViewModelProvider.Factory, body: T.() -> Unit): T {
-        val vm = ViewModelProviders.of(this, factory)[T::class.java]
+    inline fun <reified T : ViewModel> viewModel(
+        factory: ViewModelProvider.Factory,
+        body: T.() -> Unit
+    ): T {
+        val vm = ViewModelProvider(this, factory)[T::class.java]
         vm.body()
         return vm
     }
