@@ -15,7 +15,7 @@
  */
 package com.shakutara.cleanarchitecture.core.di.viewmodel
 
-import android.arch.lifecycle.ViewModel
+import androidx.lifecycle.ViewModel
 import dagger.MapKey
 import kotlin.reflect.KClass
 
