@@ -15,8 +15,8 @@
  */
 package com.shakutara.cleanarchitecture.core.platform
 
-import android.arch.lifecycle.MutableLiveData
-import android.arch.lifecycle.ViewModel
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
 import com.shakutara.cleanarchitecture.core.exception.Failure
 
 /**
