@@ -1,8 +1,10 @@
 package com.shakutara.cleanarchitecture.features.list
 
-import android.arch.lifecycle.MutableLiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.viewModelScope
 import com.shakutara.cleanarchitecture.core.interactor.UseCase
 import com.shakutara.cleanarchitecture.core.platform.BaseViewModel
+import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 class MoviesViewModel
@@ -10,9 +12,16 @@ class MoviesViewModel
 
     var movies: MutableLiveData<List<Movie>> = MutableLiveData()
 
-    fun loadMovies() = getMovieListUseCase(UseCase.None()) { it.either(::handleFailure, ::handleMovieList) }
+    fun loadMovies() = viewModelScope.launch {
+        getMovieListUseCase(UseCase.None()) {
+            it.either(
+                ::handleFailure,
+                ::handleMovieList
+            )
+        }
+    }
 
     private fun handleMovieList(movies: List<Movie>) {
-        this.movies.value = movies
+        this.movies.postValue(movies)
     }
 }
